requirements.txt
hydra-client-blueprint
flask
flask-dance[sqla]
flask-sqlalchemy
flask-login
python-dotenv
psycopg2
click
